django/requirements.txt
Django
kafka-python
confluent-kafka==2.9.0
